pyproject.toml
[project]
name = "gal_helper_backend"
version = "0.1.0"
description = "AI Agent for Galgame Support"
readme = "README.md"
requires-python = ">=3.10"
# 这里就是原 requirements.txt 的内容
dependencies = [
    # Core dependencies
    "fastapi[standard]>=0.104.0", # fastapi版本
    "uvicorn>=0.24.0", # uvicorn服务器
    "asyncpg>=0.29.0", # 异步postgres
    "aiohttp>=3.9.0", # aiohttp
    "python-dotenv>=1.0.0", # 加载.env
    "dashscope>=1.14.0", # 阿里ai版本
    "pydantic-settings>=2.12.0", # pydantic-settings
    "sqlalchemy[asyncio]>=2.0.46", # sqlalchemy ORM框架
     # postgres官方库，langchain官方指定 使用
    "psycopg>=3.3.2",

    # AI/ML
    "numpy>=1.24.0", # numpy
    "pydantic>=2.5.0", # 数据验证和解析
    # Data processing
    "openpyxl>=3.1.0", # 读写excel
    "pandas>=2.0.0", # 数据分析和数据处理库
    "python-multipart>=0.0.6", # 处理multipart/form-data
    # 文档UI
    "scalar-fastapi>=1.6.1",
    # Agent框架
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langchain-community>=0.4.1",
    "langchain-postgres>=0.0.16",
    "langgraph-checkpoint-postgres>=3.0.3",
    # uuid工具库，用于生成session_code
    "uuid6>=2025.0.1",
]

[tool.uv]
# 开发环境专用的依赖（类似 Maven 的 <scope>test</scope>）
dev-dependencies = [
    "pytest>=7.4.0",
    "ruff>=0.3.0",
    "pytest-asyncio>=0.21.0"
]
